+++ CHANGES.rst
@@ -8,3 +8,5 @@ Changelog
   [gbastien]
 - Added `@@insert-barcode` view.
   [gbastien]
+- Make tests rely on PloneMeetingTestCase.
+  [gbastien]